requirements.txt
transformers
torch
torchvision
pillow
tqdm